Rename user endpoint to profile
