+++ src/routes/index.js
@@ -4,7 +4,7 @@ const isOfficer = require("../middlewares/isOfficer");
 
 router.use(/\/auth/, require("./auth"));
 router.use("/events", require("./events"));
-router.use("/user", isAuthenticated, require("./user"));
+router.use("/profile", isAuthenticated, require("./profile"));
 router.use("/users", isAuthenticated, isOfficer, require("./users"));
 
 router.use((_req, res) => {
